Change note notation to be inspired by ABC/Helmholtz

+++ code/SoundPlayground/SoundPlayground/Parser/AbstractSyntaxTree/ContextModifiers/LengthModifierNode.cs
@@ -3,9 +3,9 @@ using SoundPlayground.Core;
 
 namespace SoundPlayground.Parser.AbstractSyntaxTree.ContextModifiers {
     public class LengthModifierNode : ContextModifierNode {
-        public int Length { get; set; }
+        public float Length { get; set; }
 
-        public LengthModifierNode ( int length ) {
+        public LengthModifierNode ( float length ) {
             Length = length;
         }
